Improved testing. Finalized in-app tracking of existing files.

+++ AndroidCSV/src/se/acoder/androidcsv/file/FileManager.java
@@ -1,5 +1,6 @@
 package se.acoder.androidcsv.file;
 
+import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 import android.content.Context;
@@ -39,22 +40,31 @@ public class FileManager {
 	private FileManager(Context c){
 		managedFiles = new LinkedList<FilePath>();
 		
-		manager_settings = new FilePath("barsys_filemanager", c);
+		manager_settings = new FilePath("filemanager", c);
 		if(FileIO.fileExists(manager_settings)){
 			List<Row> rawFilePaths = FileIO.readFile(manager_settings);
 			Log.i(TAG, "Loads managed file-paths");
 			for(Row rawFilePath : rawFilePaths ){
-				managedFiles.add(new FilePath(rawFilePath.rebuild()[0], c));
+				String[] values = rawFilePath.rebuild();
+				managedFiles.add(new FilePath(values[0], c, Extention.createExtention(values[1])));
 			}
 		}
 	}
 	
+	private Row filePathToRow(FilePath fp){
+		String[] fpSaveFormat = {fp.getSimpleName(), fp.getExtention().replace(".", "")};
+		return new Row(fpSaveFormat);
+	}
+	
 	public synchronized boolean deleteFile(FilePath fp){
-		managedFiles.remove(fp);
-		if(FileIO.fileExists(fp)){
-			Log.d(TAG,"Deletes file '"+fp.getName()+"'");
-			FileIO.deleteFile(fp);
-			return true;
+		if(managedFiles.contains(fp)){
+			managedFiles.remove(fp);
+			FileIO.removeRow(manager_settings, filePathToRow(fp));
+			if(FileIO.fileExists(fp)){
+				Log.d(TAG,"Deletes file '"+fp.getName()+"'");
+				FileIO.deleteFile(fp);
+				return true;
+			}
 		}
 		return false;
 	}
@@ -67,10 +77,31 @@ public class FileManager {
 	 */
 	public synchronized boolean appendFile(FilePath fp, List<Row> rows){
 		if(FileIO.appendFile(fp, rows)){
-			if(!FileIsManaged(fp))
+			if(!FileIsManaged(fp)){
 				managedFiles.add(fp);
+				appendFile(manager_settings, filePathToRow(fp));
+			}
 			return true;
 		}
+		return false;
+	}
+	public synchronized boolean appendFile(FilePath fp, Row row){
+		List<Row> list = new ArrayList<Row>();
+		list.add(row);
+		return appendFile(fp, list);
+	}
+	
+	/**
+	 * Mostly implemented for test reasons. Normally it is inefficient to just
+	 * fiddle around with individual rows. Better to write a temp-file and overwrite
+	 * or merge.
+	 * @param fp
+	 * @param row
+	 * @return
+	 */
+	public synchronized boolean removeRow(FilePath fp, Row row){
+		if(FileIsManaged(fp))
+			return FileIO.removeRow(fp, row);
 		return false;
 	}
 	
@@ -80,6 +111,8 @@ public class FileManager {
 	 * @return 
 	 */
 	public List<Row> readFile(FilePath fp){
+		Log.d(TAG, "File is managed: " +FileIsManaged(fp));
+		Log.d(TAG, "File exists: " +FileIO.fileExists(fp));
 		if(FileIsManaged(fp) && FileIO.fileExists(fp)){
 			return FileIO.readFile(fp);
 		}
@@ -90,14 +123,6 @@ public class FileManager {
 		if(FileIsManaged(fp) && FileIO.fileExists(fp))
 			return FileIO.saveExternal(fp);
 		return false;
-	}
-	
-	/**
-	 * Method should be called as part of shutdown procedure in order to ensure
-	 * that state-data is saved correctly.
-	 */
-	public void controlledShutdown(){
-		//appendFile(managedFiles, )
 	}
 	
 	/**
@@ -118,6 +143,8 @@ public class FileManager {
 	 * @return True if known.
 	 */
 	public static boolean FileIsManaged(FilePath fp){
+		if(fp.equals(manager_settings))
+			return true;
 		return managedFiles.contains(fp);
 	}
 }
